requirements.txt
instructor
openai
pydantic
python-dotenv